discovered speed performance problems with calling interface functions. fm effect working with workaround optimizations

+++ waves/dsp/ab_engine.h
@@ -25,7 +25,7 @@ class ABEngine: public Engine
 public:
     ABEngine();
     ~ABEngine();
-    virtual void Render(AudioDac::Frame* output, uint32_t size, uint16_t tune, uint16_t fx_amount, uint16_t fx, uint16_t morph);
+    virtual void Render(AudioDac::Frame* output, size_t size, uint16_t tune, uint16_t fx_amount, uint16_t fx, uint16_t morph);
     float Render();
     void Init();
     float GetSample(int16_t wavetable, int16_t frame, float phase);
